pyproject.toml
[project]
name = "pzh"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.15"
dependencies = [
    "mkdocs-changelog-plugin>=0.1.3",
    "mkdocs-encryptcontent-plugin>=3.1.0",
    "mkdocs-git-revision-date-localized-plugin>=1.5.1",
    "mkdocs-glightbox>=0.5.2",
    "mkdocs-heti-plugin>=0.1.6",
    "mkdocs-include-markdown-plugin>=7.2.1",
    "mkdocs-material>=9.7.1",
    "mkdocs-minify-plugin>=0.8.0",
    "mkdocs-rss-plugin>=1.17.9",
    "mkdocs-statistics-plugin>=0.1.5",
    "pyyaml>=6.0.3",
]
